Updated index, banning, submission and admin pages
Added verification for most if not all necessary configuration files (works around an annoying GitHub bug where empty files are apparently thrown out whether the uploader wants this or not...).
Fixed setting for allowing multiple active requests (the prompt was backwards...).
Implemented automatic banning based on a set of key words to compare usernames against.
Cleaned up a few settings-related issues.
Added ability to supply a "reason" for a ban. Note this reason is not yet displayed to the user, this is coming soon.
Fixed bugs in the ban list viewing pages.